Learn-English_TgBot/handlers/bot_states.py
from telebot.handler_backends import StatesGroup, State


class BotStates(StatesGroup):
    """
    Класс описывающий состояния бота.
    """
    start = State()
    play = State()
