Learn-English_TgBot/requirements.txt
annotated-types==0.7.0
certifi==2024.6.2
charset-normalizer==3.3.2
emoji==2.12.1
first==2.0.2
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
mypy==1.11.2
mypy-extensions==1.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pyTelegramBotAPI==4.20.0
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
transitions==0.9.2
typing_extensions==4.12.2
urllib3==2.2.2
